Deploy V6
arreglados los errores dependientes de la versión de java

-arreglos mínimos

+++ backend/mgb/src/main/java/com/portfolio/mgb/Service/ImpPersonaService.java
@@ -31,11 +31,11 @@ public class ImpPersonaService {
          ipersonaRepository.save(persona);
      }
      
-     public void delete(int id){
+     public void delete(long id){
          ipersonaRepository.deleteById(id);
      }
      
-     public boolean existsById(int id){
+     public boolean existsById(long id){
          return ipersonaRepository.existsById(id);
      }
